user_profile

+++ cart/views.py
@@ -1,8 +1,8 @@
 from django.shortcuts import render, redirect
 from .models import Cart
 from django.http import HttpResponse
+from django.contrib import messages
 from products.models import Product
-
  
 def cart(request):
     user = request.user
@@ -30,5 +30,9 @@ def add_cart(request, id):
 def remove(request, product_id):
     user = request.user
     cart_item = Cart.objects.filter(user=user.id)
-    cart_item.get(product_id=product_id).delete()
+    if cart_item.filter(product_id=product_id):
+        messages.success(request,'Item removed from cart successfully')
+        cart_item.get(product_id=product_id).delete()
+    else:
+        messages.error(request,'Item is not in cart')
     return redirect("cart:cart")
